Returns Json. App works with both relationships. Saving into DB works as well. Security was deleted from project

+++ src/main/java/ca/BaigiamasisBE/repositories/HelmetRepository.java
@@ -3,7 +3,7 @@ package ca.BaigiamasisBE.repositories;
 import ca.BaigiamasisBE.entities.Helmet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HelmetRepository extends JpaRepository<Helmet, Long> {
+public interface HelmetRepository extends JpaRepository<Helmet, Integer> {
 
     Helmet findById(int id);
